// frontend/packages/shared/src/utils/types/prosuctData.type.ts
export type ProductType = {
    id: string;
    name: string;
    image: string;
    price: number;
    brand: string
    discount?: number;
    description: string;
    attributes: ProductAttrType;
};

export type ProductAttrType = {
    productsId: number;
    type: string;
    category: string;
    color: string;
    size: string;
    brand: string;
    material: string; 
    countryOfOrigin: String
    weight: number
}